requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
dj-database-url==0.5.0
Django==4.0.2
django-heroku==0.3.1
djangorestframework==3.13.1
gunicorn==20.1.0
h11==0.13.0
idna==3.3
outcome==1.1.0
psycopg2==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
python-decouple==3.6
pytz==2021.3
requests==2.27.1
selenium==4.1.0
sentry-sdk==1.5.5
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.19.0
trio-websocket==0.9.2
tzdata==2021.5
urllib3==1.26.8
whitenoise==6.0.0
wsproto==1.0.0
